feat(App.tsx, package.json, tsconfig.json, babel) : Typescript configuration and setup, App.tsx updated, babel and tsconfig updated
