+++ babel.config.js
@@ -1,3 +1,16 @@
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
+  plugins: [
+    [
+      'module-resolver',
+      {
+        root: ['./src'],
+        extensions: ['.ios.js', '.android.js', '.js', '.ts', '.tsx', '.json'],
+        alias: {
+          tests: ['./tests/'],
+          '@components': './src/components',
+        },
+      },
+    ],
+  ],
 };
